requirements.txt
Pillow==11.0.0
torch==2.1.2
torchvision==0.16.2
